Remove java package management from module
Without this change, this module attempts to manage the Java package.  This
is a problem because selection of the package is a complex task when it
needs to support multiple operating systems (and versions of operating
systems).  There are already other modules that perform this task and
perform it well, which should be used instead (and managed at the profile
level).