Add newlines between entries in user.properties
Without this change, entries in the generated user.properties file are
mashed together on a single line.  This is a problem because it results in
an invalid file (or at the very least, one that doesn't come out as
intended) when more than one property is added to the file.